+++ test/Confluent.SchemaRegistry.IntegrationTests/Tests/GetLatestSchema.cs
@@ -16,6 +16,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Threading.Tasks;
 using Xunit;
 
 
@@ -24,7 +25,7 @@ namespace Confluent.SchemaRegistry.IntegrationTests
     public static partial class Tests
     {
         [Theory, MemberData(nameof(SchemaRegistryParameters))]
-        public static void GetLatestSchema(Config config)
+        public static async Task GetLatestSchema(Config config)
         {
             var topicName = Guid.NewGuid().ToString();
 
@@ -36,14 +37,14 @@ namespace Confluent.SchemaRegistry.IntegrationTests
             var sr = new CachedSchemaRegistryClient(new SchemaRegistryConfig { Url = config.Server });
 
             var subject = SubjectNameStrategy.Topic.ConstructKeySubjectName(topicName, null);
-            var id = sr.RegisterSchemaAsync(subject, testSchema1).Result;
+            var id = await sr.RegisterSchemaAsync(subject, testSchema1);
 
-            var schema = sr.GetLatestSchemaAsync(subject).Result;
+            var schema = await sr.GetLatestSchemaAsync(subject);
 
-            Assert.Equal(schema.Id, id);
-            Assert.Equal(schema.Subject, subject);
-            Assert.Equal(schema.Version, 1);
-            Assert.Equal(schema.SchemaString, testSchema1);
+            Assert.Equal(id, schema.Id);
+            Assert.Equal(subject, schema.Subject);
+            Assert.Equal(1, schema.Version);
+            Assert.Equal(testSchema1, schema.SchemaString);
         }
     }
 }